Minor fixes

+++ Assets/Scripts/Timer.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 public class Timer : MonoBehaviour
 {
     [Header("Timer settings")]
-    public float reloadTime;
-    public float currentTime; // Can be set from inspector as a start time
+    public int reloadTime;
+    public int currentTime; // Can be set from inspector as a start time
 
     private void FixedUpdate()
     {
